+++ Editor/ExtendBuildAssetBundleOptions.cs
@@ -1,0 +1,9 @@
+﻿namespace AssetBundleBrowser
+{
+    public enum ExtendBuildAssetBundleOptions
+    {
+        None = 0x0,
+        WithoutManifest = 0x1,
+        ReplaceByHash = 0x2,
+    }
+}